Adding a brew install for geturl.
geturl is a cli tool that returns a public url for any of your files.
It uploads it to your filepicker.io account and also tries to put it
on your clipboard if it can.

I had released in on HN and a couple people encouraged me to add it
to brew; as this is my first time, let me know if I did something wrong.

    Link: http://news.ycombinator.com/item?id=4598935 kkkkk